rlz/main.cpp: Pssed in the isshort option to the RLZCompress constructor.
rlz/RLZ.cpp RLZ.h:
Added the isshort option to the FactorWriter classes and changed the
code to pass it around.
Changed the way the encoding types are written to the output files to
indicate to the decompressor what sort of encodings were used. Now all
the encoding types are packed into one byte. Also changed the
FactorReader classes to be able to read the packed byte before decoding
begins.



git-svn-id: svn+ssh://nucleus-genomics.csse.unimelb.edu.au/home/genome/kuruppu/svn/src/rlz@787 3dd21c76-7d0f-4b49-bb94-8725f8c09f04