rlz/RLZ.cpp RLZ.h: Added another bit vector to keep track of the start factor for each sequence.
rlz/RLZ_index.cpp RLZ_index.h:
Added another bit vector to keep track of the start factor for each
sequence.
Implemented locate(). The search() function now takes a vector to store
the occurrences if it's invoked as a locate() query and ignores the
vector if it's invoded as a count() query.



git-svn-id: svn+ssh://nucleus-genomics.csse.unimelb.edu.au/home/genome/kuruppu/svn/src/rlz@825 3dd21c76-7d0f-4b49-bb94-8725f8c09f04